check if env exists before overwriting

+++ scripts/install.sh
@@ -4,8 +4,11 @@
 ENVFILE=".env"
 GLOBAL_DEPS=("webpack" "webpack-cli" "typescript" "ts-node")
 
-cp "$ENVFILE.template" "$ENVFILE"
-echo "Please update $ENVFILE with valid environment variables."
+# check if envfile exists
+if [ ! -f $ENVFILE ]; then
+    cp "$ENVFILE.template" "$ENVFILE"
+    echo "Please update $ENVFILE with valid environment variables."
+fi
 
 npm install # local dependencies
 npm install -g "${GLOBAL_DEPS[@]}"
